pulsi_politik_backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
click==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.2
gunicorn==21.2.0
blinker==1.9.0 
python-dotenv
pydantic 
pytest-flask